Move widgets to their own directory
