+++ lib/aixformation/aixformation_row.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:viktoriaapp/aixformation/aixformation_post.dart';
 import 'package:viktoriaapp/aixformation/aixformation_utils.dart';
 import 'package:viktoriaapp/plugins/platform/platform.dart';
-import 'package:viktoriaapp/utils/custom_row.dart';
-import 'package:viktoriaapp/utils/icons_texts.dart';
+import 'package:viktoriaapp/widgets/custom_row.dart';
+import 'package:viktoriaapp/widgets/icons_texts.dart';
 import 'package:viktoriaapp/utils/theme.dart';
 import 'package:viktoriaapp/models/models.dart';
 import 'package:transparent_image/transparent_image.dart';
