+++ lib/app/app_page.dart
@@ -9,9 +9,9 @@ import 'package:viktoriaapp/plugins/pwa/pwa.dart';
 import 'package:viktoriaapp/settings/settings_page.dart';
 import 'package:viktoriaapp/substitution_plan/substitution_plan_page.dart';
 import 'package:viktoriaapp/timetable/timetable_page.dart';
-import 'package:viktoriaapp/utils/app_bar.dart';
-import 'package:viktoriaapp/utils/custom_circular_progress_indicator.dart';
-import 'package:viktoriaapp/utils/custom_linear_progress_indicator.dart';
+import 'package:viktoriaapp/widgets/custom_app_bar.dart';
+import 'package:viktoriaapp/widgets/custom_circular_progress_indicator.dart';
+import 'package:viktoriaapp/widgets/custom_linear_progress_indicator.dart';
 import 'package:viktoriaapp/utils/notifications.dart';
 import 'package:viktoriaapp/utils/screen_sizes.dart';
 import 'package:viktoriaapp/utils/static.dart';
