+++ lib/cafetoria/cafetoria_page.dart
@@ -5,10 +5,10 @@ import 'package:url_launcher/url_launcher.dart';
 import 'package:viktoriaapp/app/app_page.dart';
 import 'package:viktoriaapp/cafetoria/cafetoria_login.dart' as dialog;
 import 'package:viktoriaapp/cafetoria/cafetoria_row.dart';
-import 'package:viktoriaapp/utils/app_bar.dart';
-import 'package:viktoriaapp/utils/empty_list.dart';
-import 'package:viktoriaapp/utils/list_group.dart';
-import 'package:viktoriaapp/utils/size_limit.dart';
+import 'package:viktoriaapp/widgets/custom_app_bar.dart';
+import 'package:viktoriaapp/widgets/empty_list.dart';
+import 'package:viktoriaapp/widgets/list_group.dart';
+import 'package:viktoriaapp/widgets/size_limit.dart';
 import 'package:viktoriaapp/utils/static.dart';
 import 'package:viktoriaapp/models/models.dart';
 import 'package:viktoriaapp/utils/theme.dart';
