+++ lib/calendar/calendar_row.dart
@@ -2,7 +2,7 @@ import 'package:add_2_calendar/add_2_calendar.dart';
 import 'package:after_layout/after_layout.dart';
 import 'package:flutter/material.dart';
 import 'package:viktoriaapp/plugins/platform/platform.dart';
-import 'package:viktoriaapp/utils/custom_row.dart';
+import 'package:viktoriaapp/widgets/custom_row.dart';
 import 'package:viktoriaapp/utils/theme.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:viktoriaapp/models/models.dart';
