+++ lib/substitution_plan/substitution_plan_page.dart
@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:timeago/timeago.dart' as timeago;
 import 'package:viktoriaapp/models/models.dart';
 import 'package:viktoriaapp/substitution_plan/substitution_plan_row.dart';
-import 'package:viktoriaapp/utils/custom_grid.dart';
-import 'package:viktoriaapp/utils/custom_hero.dart';
-import 'package:viktoriaapp/utils/empty_list.dart';
-import 'package:viktoriaapp/utils/icons_texts.dart';
-import 'package:viktoriaapp/utils/list_group.dart';
+import 'package:viktoriaapp/widgets/custom_grid.dart';
+import 'package:viktoriaapp/widgets/custom_hero.dart';
+import 'package:viktoriaapp/widgets/empty_list.dart';
+import 'package:viktoriaapp/widgets/icons_texts.dart';
+import 'package:viktoriaapp/widgets/list_group.dart';
 import 'package:viktoriaapp/utils/screen_sizes.dart';
-import 'package:viktoriaapp/utils/size_limit.dart';
+import 'package:viktoriaapp/widgets/size_limit.dart';
 import 'package:viktoriaapp/utils/static.dart';
 import 'package:viktoriaapp/utils/theme.dart';
